+++ src/datamodel/clientDataModel.js
@@ -2,7 +2,7 @@ import dataUtils from "@/utils/dataUtils";
 
 export default {
     createResource(url, id, creationDate, title, description, language, paymentMethod, keywords, version, standardlicense,
-        publisher, fileType, policyNames, ruleIds, ruleJsons, artifactId, representationId, brokerUris, samples, catalogs) {
+        publisher, fileType, policyNames, contractPeriodFromValue, contractPeriodToValue, ruleIds, ruleJsons, artifactId, representationId, brokerUris, samples, catalogs) {
         let resource = {};
         if (url === undefined) {
             resource.url = "";
@@ -68,6 +68,16 @@ export default {
             resource.policyNames = [];
         } else {
             resource.policyNames = policyNames;
+        }
+        if (contractPeriodFromValue === undefined) {
+            resource.contractPeriodFromValue = "";
+        } else {
+            resource.contractPeriodFromValue = contractPeriodFromValue.substring(0, 10);
+        }
+        if (contractPeriodToValue === undefined) {
+            resource.contractPeriodToValue = "";
+        } else {
+            resource.contractPeriodToValue = contractPeriodToValue.substring(0, 10);
         }
         if (ruleIds === undefined) {
             resource.ruleIds = [];
@@ -313,7 +323,7 @@ export default {
         return this.createApp(id, idsApp.title, idsApp.description, "APP");
     },
 
-    convertIdsResource(idsResource, representation, policyNames, ruleIds, ruleJsons, artifactId, brokerUris, catalogs) {
+    convertIdsResource(idsResource, representation, policyNames, contractPeriodFromValue, contractPeriodToValue, ruleIds, ruleJsons, artifactId, brokerUris, catalogs) {
         let title = idsResource.title;
         if (title.includes("\"@en")) {
             title = idsResource.title.substring(1, idsResource.title.lastIndexOf("\""));
@@ -331,7 +341,7 @@ export default {
 
         return this.createResource(idsResource._links.self.href, dataUtils.getIdOfConnectorResponse(idsResource), idsResource.creationDate, title, description,
             idsResource.language.replace("https://w3id.org/idsa/code/", ""), idsResource.paymentModality, idsResource.keywords,
-            idsResource.version, idsResource.license, idsResource.publisher, fileType, policyNames, ruleIds, ruleJsons, artifactId, representationId, brokerUris,
+            idsResource.version, idsResource.license, idsResource.publisher, fileType, policyNames, contractPeriodFromValue, contractPeriodToValue, ruleIds, ruleJsons, artifactId, representationId, brokerUris,
             idsResource.samples, catalogs);
     },