+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Filter offered resources by catalog
+- Set contract validity period for offered resources
 
 ## [8.6.0] - 2021-10-08